+++ arch/arc/core/prep_c.c
@@ -22,8 +22,8 @@
 #include <zephyr/arch/arc/v2/aux_regs.h>
 #include <zephyr/arch/arc/cluster.h>
 #include <zephyr/kernel_structs.h>
-#include <kernel_internal.h>
 #include <zephyr/arch/common/xip.h>
+#include <zephyr/arch/common/init.h>
 #include <zephyr/platform/hooks.h>
 #include <zephyr/arch/cache.h>
