+++ arch/arm/core/cortex_m/prep_c.c
@@ -24,6 +24,7 @@
 #include <zephyr/arch/cache.h>
 #include <cortex_m/debug.h>
 #include <zephyr/arch/common/xip.h>
+#include <zephyr/arch/common/init.h>
 
 /*
  * GCC can detect if memcpy is passed a NULL argument, however one of
