+++ arch/arm64/core/prep_c.c
@@ -14,11 +14,11 @@
  * initialization is performed.
  */
 
-#include <kernel_internal.h>
 #include <zephyr/linker/linker-defs.h>
 #include <zephyr/platform/hooks.h>
 #include <zephyr/arch/cache.h>
 #include <zephyr/arch/common/xip.h>
+#include <zephyr/arch/common/init.h>
 
 extern void z_arm64_mm_init(bool is_primary_core);
