+++ arch/common/CMakeLists.txt
@@ -17,6 +17,7 @@ if(CONFIG_GEN_ISR_TABLES)
   )
 endif()
 
+zephyr_library_sources(init.c)
 zephyr_library_sources_ifdef(CONFIG_XIP xip.c)
 
 zephyr_library_sources_ifdef(
