+++ arch/mips/core/prep_c.c
@@ -9,10 +9,11 @@
  * @brief Full C support initialization
  */
 
-#include <kernel_internal.h>
 #include <zephyr/irq.h>
 #include <zephyr/platform/hooks.h>
 #include <zephyr/arch/cache.h>
+#include <zephyr/arch/common/xip.h>
+#include <zephyr/arch/common/init.h>
 
 static void interrupt_init(void)
 {
