+++ arch/rx/core/prep_c.c
@@ -15,12 +15,12 @@
  * initialization is performed.
  */
 
-#include <kernel_internal.h>
 #include <zephyr/kernel.h>
 #include <zephyr/logging/log.h>
 #include <zephyr/toolchain.h>
 #include <zephyr/linker/sections.h>
 #include <zephyr/arch/common/xip.h>
+#include <zephyr/arch/common/init.h>
 
 K_KERNEL_PINNED_STACK_ARRAY_DEFINE(z_initialization_process_stacks, CONFIG_MP_MAX_NUM_CPUS,
 				   CONFIG_INITIALIZATION_STACK_SIZE);
