+++ arch/x86/core/intel64/cpu.c
@@ -15,6 +15,7 @@
 #include <zephyr/drivers/interrupt_controller/loapic.h>
 #include <cet.h>
 #include <zephyr/arch/common/xip.h>
+#include <zephyr/arch/common/init.h>
 #ifdef CONFIG_ACPI
 #include <zephyr/arch/x86/cpuid.h>
 #include <zephyr/acpi/acpi.h>
