+++ arch/x86/core/prep_c.c
@@ -5,14 +5,16 @@
  */
 
 #include <zephyr/kernel.h>
-#include <kernel_internal.h>
+#include <kernel_arch_func.h>
+#include <kernel_arch_interface.h>
 #include <zephyr/arch/x86/multiboot.h>
 #include <zephyr/arch/x86/efi.h>
 #include <x86_mmu.h>
 #include <zephyr/platform/hooks.h>
-#include <zephyr/arch/cache.h>
+#include <zephyr/cache.h>
+#include <zephyr/arch/common/init.h>
 
-extern FUNC_NORETURN void z_cstart(void);
+
 extern void x86_64_irq_init(void);
 
 #if !defined(CONFIG_X86_64)
