+++ kernel/init.c
@@ -40,6 +40,8 @@
 #include <zephyr/logging/log.h>
 #include <zephyr/pm/device_runtime.h>
 #include <zephyr/internal/syscall_handler.h>
+#include <zephyr/arch/common/init.h>
+
 LOG_MODULE_REGISTER(os, CONFIG_KERNEL_LOG_LEVEL);
 
 /* the only struct z_kernel instance */
@@ -177,122 +179,6 @@ static struct k_obj_core_stats_desc  kernel_stats_desc = {
 };
 #endif /* CONFIG_OBJ_CORE_STATS_SYSTEM */
 #endif /* CONFIG_OBJ_CORE_SYSTEM */
-
-/* LCOV_EXCL_START
- *
- * This code is called so early in the boot process that code coverage
- * doesn't work properly. In addition, not all arches call this code,
- * some like x86 do this with optimized assembly
- */
-
-/**
- * @brief equivalent of memset() for early boot usage
- *
- * Architectures that can't safely use the regular (optimized) memset very
- * early during boot because e.g. hardware isn't yet sufficiently initialized
- * may override this with their own safe implementation.
- */
-__boot_func
-void __weak z_early_memset(void *dst, int c, size_t n)
-{
-	(void) memset(dst, c, n);
-}
-
-/**
- * @brief equivalent of memcpy() for early boot usage
- *
- * Architectures that can't safely use the regular (optimized) memcpy very
- * early during boot because e.g. hardware isn't yet sufficiently initialized
- * may override this with their own safe implementation.
- */
-__boot_func
-void __weak z_early_memcpy(void *dst, const void *src, size_t n)
-{
-	(void) memcpy(dst, src, n);
-}
-
-/**
- * @brief Clear BSS
- *
- * This routine clears the BSS region, so all bytes are 0.
- */
-__boot_func
-void z_bss_zero(void)
-{
-	if (IS_ENABLED(CONFIG_SKIP_BSS_CLEAR)) {
-		return;
-	}
-
-	z_early_memset(__bss_start, 0, __bss_end - __bss_start);
-#if DT_NODE_HAS_STATUS_OKAY(DT_CHOSEN(zephyr_ccm))
-	z_early_memset(&__ccm_bss_start, 0,
-		       (uintptr_t) &__ccm_bss_end
-		       - (uintptr_t) &__ccm_bss_start);
-#endif
-#if DT_NODE_HAS_STATUS_OKAY(DT_CHOSEN(zephyr_dtcm))
-	z_early_memset(&__dtcm_bss_start, 0,
-		       (uintptr_t) &__dtcm_bss_end
-		       - (uintptr_t) &__dtcm_bss_start);
-#endif
-#if DT_NODE_HAS_STATUS_OKAY(DT_CHOSEN(zephyr_ocm))
-	z_early_memset(&__ocm_bss_start, 0,
-		       (uintptr_t) &__ocm_bss_end
-		       - (uintptr_t) &__ocm_bss_start);
-#endif
-#ifdef CONFIG_CODE_DATA_RELOCATION
-	extern void bss_zeroing_relocation(void);
-
-	bss_zeroing_relocation();
-#endif	/* CONFIG_CODE_DATA_RELOCATION */
-#ifdef CONFIG_COVERAGE_GCOV
-	z_early_memset(&__gcov_bss_start, 0,
-		       ((uintptr_t) &__gcov_bss_end - (uintptr_t) &__gcov_bss_start));
-#endif /* CONFIG_COVERAGE_GCOV */
-#ifdef CONFIG_NOCACHE_MEMORY
-	z_early_memset(&_nocache_ram_start, 0,
-		       (uintptr_t)&_nocache_ram_end - (uintptr_t)&_nocache_ram_start);
-#endif
-}
-
-#ifdef CONFIG_LINKER_USE_BOOT_SECTION
-/**
- * @brief Clear BSS within the boot region
- *
- * This routine clears the BSS within the boot region.
- * This is separate from z_bss_zero() as boot region may
- * contain symbols required for the boot process before
- * paging is initialized.
- */
-__boot_func
-void z_bss_zero_boot(void)
-{
-	z_early_memset(&lnkr_boot_bss_start, 0,
-		       (uintptr_t)&lnkr_boot_bss_end
-		       - (uintptr_t)&lnkr_boot_bss_start);
-}
-#endif /* CONFIG_LINKER_USE_BOOT_SECTION */
-
-#ifdef CONFIG_LINKER_USE_PINNED_SECTION
-/**
- * @brief Clear BSS within the pinned region
- *
- * This routine clears the BSS within the pinned region.
- * This is separate from z_bss_zero() as pinned region may
- * contain symbols required for the boot process before
- * paging is initialized.
- */
-#ifdef CONFIG_LINKER_USE_BOOT_SECTION
-__boot_func
-#else
-__pinned_func
-#endif /* CONFIG_LINKER_USE_BOOT_SECTION */
-void z_bss_zero_pinned(void)
-{
-	z_early_memset(&lnkr_pinned_bss_start, 0,
-		       (uintptr_t)&lnkr_pinned_bss_end
-		       - (uintptr_t)&lnkr_pinned_bss_start);
-}
-#endif /* CONFIG_LINKER_USE_PINNED_SECTION */
 
 #ifdef CONFIG_REQUIRES_STACK_CANARIES
 #ifdef CONFIG_STACK_CANARIES_TLS
